// src/shared/interfaces/parent-notifications-types.enum.ts
import { registerEnumType } from '@nestjs/graphql';

export enum ParentNotificationsTypes {
  HABIT_CREATION_REQUESTED = 'HABIT_CREATION_REQUESTED',
  HABIT_UPDATE_REQUESTED = 'HABIT_UPDATE_REQUESTED',
  AWARD_CREATION_REQUESTED = 'AWARD_CREATION_REQUESTED',
  AWARD_UPDATE_REQUESTED = 'AWARD_UPDATE_REQUESTED',
  AWARD_PURCHASED = 'AWARD_PURCHASED',
  HABIT_DELETION_REQUESTED = 'HABIT_DELETION_REQUESTED',
  CHILD_ACTIVATED = 'CHILD_ACTIVATED',
}

registerEnumType(ParentNotificationsTypes, {
  name: 'NotificationsTypes',
});